+++ packages/buble/types/index.d.ts
@@ -4,7 +4,7 @@ import { Plugin } from 'rollup';
 
 export interface RollupBubleOptions extends TransformOptions {
   /**
-   * A minimatch pattern, or array of patterns, of files that should be
+   * A picomatch pattern, or array of patterns, of files that should be
    * processed by this plugin (if omitted, all files are included by default)
    */
   include?: FilterPattern;