+++ packages/node-resolve/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @rollup/plugin-node-resolve ChangeLog
 
+## v14.1.0
+
+_2022-09-12_
+
+### Features
+
+- feat: add new option, modulePaths (#1104)
+
+## v14.0.1
+
+_2022-09-08_
+
+### Bugfixes
+
+- fix: handle circular commonjs (#1259)
+
+## v14.0.0
+
+_2022-09-06_
+
+### Breaking Changes
+
+- fix: preserve moduleSideEffects when re-resolving files (#1245)
+
 ## v13.3.0
 
 _2022-05-02_